Kmc_space/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kmc-env"
version = "0.1.0"
description = "KMC - Key Management & Coordination environment for aligned intelligence (OpenEnv)"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
demo = ["gradio>=4.0"]
dev = ["pytest>=8.0.0"]

[project.scripts]
kmc-server = "kmc_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["kmc_env", "kmc_env.server"]
package-dir = {"kmc_env" = ".", "kmc_env.server" = "server"}